+++ srcs/AABB.cpp
@@ -10,8 +10,8 @@
 // Constructs the AABB with default values
 AABB::AABB(void)
 {
-    this->_minimum = Vector3(0.0f, 0.0f, 0.0f);
-    this->_minimum = Vector3(0.0f, 0.0f, 0.0f);
+    this->_minimum = Vector3(0.0, 0.0, 0.0);
+    this->_minimum = Vector3(0.0, 0.0, 0.0);
 }
 
 // Constructs the AABB with custom values
@@ -34,19 +34,19 @@ Vector3 AABB::getMaximum(void) const
 }
 
 // Returns true if 'ray' intersects with the AABB
-bool    AABB::hit(Ray& ray, float t_max) const
+bool    AABB::hit(Ray& ray, double t_max) const
 {
     for (int a = 0; a < 3; a++)
     {
-        float invD = 1.0f / ray.getDirection()[a];
-        float t0 = (this->_minimum[a] - ray.getOrigin()[a]) * invD;
-        float t1 = (this->_maximum[a] - ray.getOrigin()[a]) * invD;
-        if (invD < 0.0f)
+        double invD = 1.0 / ray.getDirection()[a];
+        double t0 = (this->_minimum[a] - ray.getOrigin()[a]) * invD;
+        double t1 = (this->_maximum[a] - ray.getOrigin()[a]) * invD;
+        if (invD < 0.0)
         {
             std::swap(t0, t1);
         }
 
-        float t_min = t0 > T_MIN ? t0 : T_MIN;
+        double t_min = t0 > T_MIN ? t0 : T_MIN;
         t_max = t1 < t_max ? t1 : t_max;
         if (t_max <= t_min)
         {
