+++ srcs/Camera.cpp
@@ -7,14 +7,14 @@
 // Constructs the Camera with default values
 Camera::Camera(void)
 {
-	this->_lookFrom = Vector3(0.0f, 0.0f, 15.0f);
-	this->_lookAt = Vector3(0.0f, 0.0f, 0.0f);
+	this->_lookFrom = Vector3(0.0, 0.0, 15.0);
+	this->_lookAt = Vector3(0.0, 0.0, 0.0);
 	this->_fov = 65;
-	this->_aperture = 2.0f;
+	this->_aperture = 2.0;
 }
 
 // Constructs the Camera with custom values
-Camera::Camera(Vector3 lookFrom, Vector3 lookAt, short fov, float aperture)
+Camera::Camera(Vector3 lookFrom, Vector3 lookAt, short fov, double aperture)
 {
 	this->_lookFrom = lookFrom;
 	this->_lookAt = lookAt;
@@ -39,7 +39,7 @@ short		Camera::getFOV(void) const
 }
 
 // Returns the Camera's Aperture
-float	Camera::getAperture(void) const
+double	Camera::getAperture(void) const
 {
 	return (this->_aperture);
 }
