+++ srcs/Color.cpp
@@ -8,13 +8,13 @@
 // Constructs the Color with default values
 Color::Color(void)
 {
-	this->_red = 0.0f;
-	this->_green = 0.0f;
-	this->_blue = 0.0f;
+	this->_red = 0.0;
+	this->_green = 0.0;
+	this->_blue = 0.0;
 }
 
-// Constructs the Color with custom float values [ 0.0 => 1.0 ]
-Color::Color(float r, float g, float b)
+// Constructs the Color with custom double values [ 0.0 => 1.0 ]
+Color::Color(double r, double g, double b)
 {
 	this->_red = r;
 	this->_green = g;
@@ -22,37 +22,37 @@ Color::Color(float r, float g, float b)
 }
 
 // Returns the Color's Red value
-float	Color::getRed(void) const
+double	Color::getRed(void) const
 {
 	return (this->_red);
 }
 
 // Sets the Color's Red value
-void	Color::setRed(float red)
+void	Color::setRed(double red)
 {
 	this->_red = red;
 }
 
 // Returns the Color's Green value
-float	Color::getGreen(void) const
+double	Color::getGreen(void) const
 {
 	return (this->_green);
 }
 
 // Sets the Color's Green value
-void	Color::setGreen(float green)
+void	Color::setGreen(double green)
 {
 	this->_green = green;
 }
 
 // Returns the Color's Blue value
-float	Color::getBlue(void) const
+double	Color::getBlue(void) const
 {
 	return (this->_blue);
 }
 
 // Sets the Color's Blue value
-void	Color::setBlue(float blue)
+void	Color::setBlue(double blue)
 {
 	this->_blue = blue;
 }
@@ -65,7 +65,7 @@ Color&	Color::operator+=(const Color &color2)
 	return (*this);
 }
 
-Color&	Color::operator/=(const float f)
+Color&	Color::operator/=(const double f)
 {
 	this->_red /= f;
 	this->_green /= f;
@@ -73,7 +73,7 @@ Color&	Color::operator/=(const float f)
 	return (*this);
 }
 
-Color	Color::operator*(const float f) const
+Color	Color::operator*(const double f) const
 {
 	return (Color(this->_red * f, this->_green * f, this->_blue * f));
 }
@@ -83,7 +83,7 @@ Color	Color::operator*(const Color color) const
 	return (Color(this->_red * color.getRed(), this->_green * color.getGreen(), this->_blue * color.getBlue()));
 }
 
-Color	Color::operator/(const float f) const
+Color	Color::operator/(const double f) const
 {
 	return (Color(this->_red / f, this->_green / f, this->_blue / f));
 }
